Return a message instead of void from rpc udf
Pull Request resolved: https://github.com/pytorch/pytorch/pull/25283

Return a message instead of void from rpc udf
This is to help thrift style rpc where there is no need for explicit send for a response.
We need to figure out how to solve the non-blocking callback case but don't want to block the thrift backed rpc agent implementation till then.
ghstack-source-id: 89130305

Differential Revision: [D16825072](https://our.internmc.facebook.com/intern/diff/D16825072/)