Implement FoldConvBatchnorm2d pass.
For now it will be used in quantization, but it can be used as a
standalone pass too.

Couple of things are not finished at this moment:
- Batchnorm.eps value is hardcoded. This is bad and wrong, but we cannot
access fields listed in __constants__ from IR now. Once we fix this, we
should remove the hardcoded value.
- We do not remove Batchnorm submodules from the parent module even when
they were merged into a Conv. Once we figure out API for removing
attributes and modules, we should fix this.

ghstack-source-id: 0202e07ca01cc02da2160c318911ede9286903b8
Pull Request resolved: https://github.com/pytorch/pytorch/pull/25282