Make scatter/gather arguments optional
For both scatter and gather, only the source and destination rank,
respectively, need to supply a list of tensors. The `scatter_list` and
`gather_list` arguments were mandatory, however, and this has resulted
in some confusion. This commit makes both the `scatter_list` and
`gather_list`, and the `src` and `dst` arguments optional.

Closes #25463.

ghstack-source-id: 4ea57fa167948509d96e67510075ae2887128aec
Pull Request resolved: https://github.com/pytorch/pytorch/pull/25575