Fix named tensor printing
There are a number of ops that get called while printing tensors
depending on how large the tensors are. This PR makes it so that before
we attempt to format tensor data for printing, we drop the names of the
tensor (if there are any). This is easier than supporting named tensors
for all of those ops (which should happen eventually).

Test Plan:
- new test [namedtensor ci]

ghstack-source-id: a71c5316d231345b94c1a7dffac41767ceba99e4
Pull Request resolved: https://github.com/pytorch/pytorch/pull/25564