Fix clang-tidy script
The clang-tidy driver script generates a chunk whitelist per file so
that it only shows errors for lines that were actually changed. If a
change removes the chunk the count is equal to 0. If the chunk happens
to be at the start of the file, and the start position is equal 0,
clang-tidy fails to run. This change filters out those chunks.

ghstack-source-id: dac5352b5643c6988db96825130d5192db70ccf6
Pull Request resolved: https://github.com/pytorch/pytorch/pull/25652