Rename tensor.view_names -> tensor.renamed
This function renames the dimensions of a tensor out-of-place. Because
of that, I think `tensor.renamed(...)` is a clearer name: `view_names`
has the connotation that we can use names to `view` our tensors with a
"different shape", but what this function really does is let us rename a
tensor no matter the previous names.

`tensor.names_`, the in-place version of this, is unchanged for now.
However, we might delete this or not advertise it if it has no use case
and also because its naming is a little inconsistent with `tensor.renamed`.

Test Plan:
- [namedtensor ci]

ghstack-source-id: 27477da7cb405f907b52196850c66aa15e155096
Pull Request resolved: https://github.com/pytorch/pytorch/pull/25711