Only default USE_DISTRIBUTED=True on Linux
Pull Request resolved: https://github.com/pytorch/pytorch/pull/25725

After landing #25260 the macOS wheel builds started to fail. It turns
out that if not specified, the setup helpers default USE_DISTRIBUTED
to true on all platforms except Windows.

This commit updates that such that USE_DISTRIBUTED only defaults to
true on Linux. More work is needed to enable it by default on macOS.

[test wheel]

ghstack-source-id: 89571701

Differential Revision: [D17211695](https://our.internmc.facebook.com/intern/diff/D17211695/)