Make arguments of Module::dump easier to remember.
Summary:
Previously we had `omit_method_bodies`, `omit_attr_values` and
`omit_param_values`. They were called the same in the python bindings
and it was hard to remember their proper spelling. This PR changes them
to `code`, `attrs`, and `params` which are might easier to remember. It
also flips their meaning - now they enable printing instead of disabling
it. I also changed the default values to 'print all' from 'print
nothing', as that's the most usual way of using it.

ghstack-source-id: f05fa36a5cad0b5746ea1bad5b3d8f10dc7de41a
Pull Request resolved: https://github.com/pytorch/pytorch/pull/25740