Add helper to aggregate multiple process groups
The round robin process group can be constructed from multiple other
process groups. Every collective call against this new process group
is delegated to the specified process groups in a round robin fashion.

Doing so may benefit performance when calling into multiple NCCL
process groups. Instead of adding support for round-robin usage of
NCCL communicators, we achieve the same without changing the NCCL
process group and adding this wrapper class.

The API to create this round robin process group is a bit harsh. If we
find it adds significant benefit we can revisit and make this a first
class citizen in the torch.distributed module.

Differential Revision: [D17226323](https://our.internmc.facebook.com/intern/diff/D17226323/)

ghstack-source-id: 89618924
Pull Request resolved: https://github.com/pytorch/pytorch/pull/25768