Remove superfluous check for POLLIN in TCPStore
The check is practically equivalent to checking for equivalence with
POLLIN (because the constant is a single bit and poll(2) is asked to
check for POLLIN). On macOS, if a client disconnects, POLLHUP will be
set as well, and the check fails. Instead of performing the check and
letting it fail, we can simply run the `query` function and catch
exceptions, in case we see EOF.

ghstack-source-id: 3f2d4a7e567279ceb18f85bd960665fbbe59a39b
Pull Request resolved: https://github.com/pytorch/pytorch/pull/25911