[jit] Refactor emitIsInstance
This is in preparation for actually emitting a dynamic isinstance check instruction.
It re-arranges the  logic so that all the types and properties to check
against are in a flat list. In the future this flat list will be encoded
into an actual instruction if we determine that we cannot perform
the check statically.

gh-metadata: pytorch pytorch 26061 gh/zdevito/106/head