Implement tensor.align_as(other), change tensor.align_to(names)
`tensor.align_to(*names)` permutes the dimensions of `tensor` and adds
additional 1-sized dimensions such that the output tensor has dimensions
in the same order as `names`. All dimensions of `tensor` must be
present in `names`, in addition, this function requires that all dims of
`tensor` be named.

`tensor.align_as(other)` is equivalent to
`tensor.align_to(*other.names)`.

I'm planning on changing `torch.align_tensors(*tensors)` to align closer
to these semantics because there didn't seem to be a clear use case for the old
semantics that preserve unnamed dimensions. That will come in a future
change.

Test Plan:
- new tests [namedtensor ci]

ghstack-source-id: 98a15ab5883becb8e9775f771e7872eb1ffab281
Pull Request resolved: https://github.com/pytorch/pytorch/pull/25843